wsi-backend/requirements.txt
fastapi==0.104.0
Pillow== 10.1.0

# This is necessary to upload form data to fastapi
python-multipart==0.0.6

# Web server
uvicorn[standard]==0.23.2

# Testing
pytest==7.4.2

# Documentation
pdoc3==0.10.0